+++ tests/test_as_people.py
@@ -6,7 +6,7 @@ import asyncio
 class TestPeople(TestCaseWithLog):
 
     @TestCaseWithLog.async_test
-    async def test_001(self):
+    async def test_001_all_details(self):
         me = await self.async_api.people.me()
         users = [u async for u in self.async_api.people.list_gen()]
         # try to provoke 429