+++ wxc_sdk/people/__init__.py
@@ -12,7 +12,6 @@ To learn more about managing people in a room see the Memberships API.
 """
 
 import datetime
-import json
 from collections.abc import Generator
 from enum import Enum
 from typing import Optional
@@ -20,7 +19,7 @@ from typing import Optional
 from pydantic import Field
 
 from ..api_child import ApiChild
-from ..base import ApiModel, to_camel, webex_id_to_uuid
+from ..base import ApiModel, to_camel, webex_id_to_uuid, ApiModelWithErrors
 
 __all__ = ['PhoneNumberType', 'PhoneNumber', 'SipType', 'SipAddress', 'PeopleStatus', 'PersonType', 'Person',
            'PeopleApi']
@@ -85,14 +84,14 @@ class PersonType(str, Enum):
     app_user = 'appuser'  #: account is a guest user
 
 
-class Person(ApiModel):
+class Person(ApiModelWithErrors):
     """
     Webex person
     """
     #: A unique identifier for the person.
-    person_id: str = Field(alias='id')
+    person_id: Optional[str] = Field(alias='id')
     #: The email addresses of the person.
-    emails: list[str]
+    emails: Optional[list[str]]
     #: Phone numbers for the person.
     phone_numbers: Optional[list[PhoneNumber]]
     #: The Webex Calling extension for the person. Only applies to a person with a Webex Calling license
@@ -111,15 +110,15 @@ class Person(ApiModel):
     #: The URL to the person's avatar in PNG format
     avatar: Optional[str]
     #: The ID of the organization to which this person belongs.
-    org_id: str
+    org_id: Optional[str]
     #: An array of role strings representing the roles to which this person belongs.
     roles: Optional[list[str]]
     #: An array of license strings allocated to this person.
     licenses: Optional[list[str]]
     #: The date and time the person was created.
-    created: datetime.datetime
+    created: Optional[datetime.datetime]
     #: The date and time the person was last changed.
-    last_modified: str
+    last_modified: Optional[datetime.datetime]
     #: The time zone of the person if configured. If no timezone is configured on the account, this field will not be
     #: present
     timezone: Optional[str]
@@ -148,7 +147,7 @@ class Person(ApiModel):
     #: False: the person cannot log into Webex
     login_enabled: Optional[bool]
     #: The type of person account, such as person or bot.
-    person_type: PersonType = Field(alias='type')
+    person_type: Optional[PersonType] = Field(alias='type')
 
     @property
     def person_id_uuid(self) -> str:
@@ -228,6 +227,46 @@ class PeopleApi(ApiChild, base='people'):
         # noinspection PyTypeChecker
         return self.session.follow_pagination(url=ep, model=Person, params=params)
 
+    def create(self, settings: Person, calling_data: bool = False) -> Person:
+        """
+        Create a Person
+
+        Create a new user account for a given organization. Only an admin can create a new user account.
+
+        At least one of the following body parameters is required to create a new user: displayName, firstName,
+        lastName.
+
+        Currently, users may have only one email address associated with their account. The emails parameter is an
+        array, which accepts multiple values to allow for future expansion, but currently only one email address will
+        be used for the new user.
+
+        Admin users can include Webex calling (BroadCloud) user details in the response by specifying callingData
+        parameter as true.
+
+        When doing attendee management, to make the new user an attendee for a site: append #attendee to the siteUrl
+        parameter (eg: mysite.webex.com#attendee).
+
+        :param settings: settings for new user
+        :type settings: Person
+        :param calling_data: Include Webex Calling user details in the response.
+        :type calling_data: bool
+        :return: new user
+        :rtype: Person
+        """
+        params = calling_data and {'callingData': 'true'} or None
+        url = self.ep()
+        data = settings.json(exclude={'person_id': True,
+                                      'created': True,
+                                      'last_modified': True,
+                                      'timezone': True,
+                                      'last_activity': True,
+                                      'sip_addresses': True,
+                                      'status': True,
+                                      'invite_pending': True,
+                                      'login_enabled': True,
+                                      'person_type': True})
+        return Person.parse_obj(self.post(url, data=data, params=params))
+
     def details(self, person_id: str, calling_data: bool = False) -> Person:
         """
         Shows details for a person, by ID.
@@ -293,19 +332,24 @@ class PeopleApi(ApiChild, base='people'):
         :return: Person details
         :rtype: Person
         """
-        params = {}
-        if calling_data:
-            params['callingData'] = 'true'
-        if show_all_types:
-            params['showAllTypes'] = 'true'
+        params = calling_data and {'callingData': 'true'} or None
 
-        # set of attributes that can be updated and should be included in body of PUT
-        can_update = {'emails', 'phoneNumbers', 'extension', 'locationId', 'displayName', 'firstName', 'lastName',
-                      'nickName', 'avatar', 'orgId', 'roles', 'licenses', 'siteUrls', 'loginEnabled'}
-        data = {k: v for k, v in json.loads(person.json()).items()
-                if k in can_update}
+        if not all(v is not None
+                   for v in (person.display_name, person.first_name, person.last_name)):
+            raise ValueError('display_name, first_name, and last_name are required')
+
+        # some attributes should not be included in update
+        data = person.json(exclude={'created': True,
+                                    'last_modified': True,
+                                    'timezone': True,
+                                    'last_activity': True,
+                                    'sip_addresses': True,
+                                    'status': True,
+                                    'invite_pending': True,
+                                    'login_enabled': True,
+                                    'person_type': True})
         ep = self.ep(path=person.person_id)
-        return Person.parse_obj(self.put(url=ep, json=data))
+        return Person.parse_obj(self.put(url=ep, data=data, params=params))
 
     def me(self, calling_data: bool = False) -> Person:
         """